feat: serve public static files
